Switch project to TypeScript
