+++ Dockerfile
@@ -6,13 +6,16 @@ WORKDIR /usr/src/app
 
 # Install dependencies
 COPY package*.json ./
-RUN npm install --production
+RUN npm install
 
 # Copy application source
 COPY . .
+
+# Build TypeScript
+RUN npm run build && npm prune --production
 
 # Expose port
 EXPOSE 3000
 
 # Start the server
-CMD ["node", "index.js"]
+CMD ["node", "dist/index.js"]
