--- index.js
@@ -1,13 +1,0 @@
-const express = require('express');
-
-const app = express();
-
-const PORT = 3000;
-
-app.get('/', (req, res) => {
-    res.send('Hello, world!');
-});
-
-app.listen(PORT, () => {
-    console.log(`Server is running on port ${PORT}`);
-});
